Fixed compilation of the packet_debug extension
Reviewed-on: http://review.couchbase.org/9051
Tested-by: Trond Norbye <trond.norbye@gmail.com>

+++ src/packet_debug.c
@@ -113,9 +113,9 @@ static void debug_tap_connect_request(protocol_binary_request_header *req)
         return ;
     }
 
-    int nextbyte = sizeof(*req);
+    size_t nextbyte = sizeof(*req);
     protocol_binary_request_tap_connect *t = (void*)req;
-    printf("Flags        (%02d-%02d): ",
+    printf("Flags        (%02zu-%02zu): ",
            nextbyte, nextbyte + (int)sizeof(t->message.body.flags) - 1);
     print_uint32(t->message.body.flags, false, true);
     nextbyte += sizeof(t->message.body.flags);
@@ -155,7 +155,7 @@ static void debug_tap_connect_request(protocol_binary_request_header *req)
     uint8_t *ptr = (void*)(t->bytes + sizeof(t->bytes));
     uint16_t keylen = ntohs(req->request.keylen);
     if (keylen > 0) {
-        printf("Name         (%02d-%02d): [", nextbyte, nextbyte + keylen - 1);
+        printf("Name         (%02zu-%02zu): [", nextbyte, nextbyte + keylen - 1);
         print_string((void*)ptr, (size_t)keylen);
         printf("]\n");
         nextbyte += keylen;
@@ -165,7 +165,7 @@ static void debug_tap_connect_request(protocol_binary_request_header *req)
     if (flags & TAP_CONNECT_FLAG_BACKFILL) {
         uint64_t backfill;
         memcpy(&backfill, ptr, sizeof(backfill));
-        printf("Backfill date(%02d-%02d): ",
+        printf("Backfill date(%02zu-%02zu): ",
                nextbyte, nextbyte + (int)sizeof(backfill) - 1);
         print_uint64(backfill, true, true);
         nextbyte += sizeof(backfill);
@@ -177,10 +177,10 @@ static void debug_tap_connect_request(protocol_binary_request_header *req)
         uint16_t val;
         memcpy(&val, ptr, sizeof(val));
         num = ntohs(val);
-        printf("VBucket list (%02d-%02d): \n",
-               nextbyte, nextbyte + (2 * num) + 1);
+        printf("VBucket list (%02zu-%02zu): \n",
+               nextbyte, nextbyte + (2U * num) + 1);
 
-        printf("    # listed (%02d-%02d): ",
+        printf("    # listed (%02zu-%02zu): ",
                nextbyte, nextbyte + (int)sizeof(num) - 1);
 
         print_uint16(val, true, true);
@@ -189,7 +189,7 @@ static void debug_tap_connect_request(protocol_binary_request_header *req)
         for (uint16_t ii = 0; ii < num; ++ii) {
             memcpy(&val, ptr, sizeof(num));
             if (ii < 5 || num < 10) {
-                printf("    vbucket  (%02d-%02d): ",
+                printf("    vbucket  (%02zu-%02zu): ",
                        nextbyte, nextbyte + (int)sizeof(val) - 1);
                 print_uint16(val, true, true);
             } else if (ii == num - 1) {
@@ -205,14 +205,14 @@ static void debug_tap_connect_request(protocol_binary_request_header *req)
 // @todo fix engine-specific bytes!!
 
 static void debug_tap_common_body(protocol_binary_request_tap_no_extras *req) {
-    int nextbyte = sizeof(req->message.header);
-    printf("Engine priv. (%02d-%02d): ",
+    size_t nextbyte = sizeof(req->message.header);
+    printf("Engine priv. (%02zu-%02zu): ",
            nextbyte,
            nextbyte + (int)sizeof(req->message.body.tap.enginespecific_length) - 1);
     print_uint16(req->message.body.tap.enginespecific_length, true, true);
     nextbyte += sizeof(req->message.body.tap.enginespecific_length);
 
-    printf("Flags        (%02d-%02d): ",
+    printf("Flags        (%02zu-%02zu): ",
            nextbyte,
            nextbyte + (int)sizeof(req->message.body.tap.flags) - 1);
     print_uint16(req->message.body.tap.flags, false, true);
@@ -236,16 +236,16 @@ static void debug_tap_common_body(protocol_binary_request_tap_no_extras *req) {
         }
     }
 
-    printf("TTL             (%02d): %02x\n", nextbyte,
+    printf("TTL             (%02zu): %02x\n", nextbyte,
            req->message.body.tap.ttl);
     ++nextbyte;
-    printf("Reserved        (%02d): %02x\n", nextbyte,
+    printf("Reserved        (%02zu): %02x\n", nextbyte,
            req->message.body.tap.res1);
     ++nextbyte;
-    printf("Reserved        (%02d): %02x\n", nextbyte,
+    printf("Reserved        (%02zu): %02x\n", nextbyte,
            req->message.body.tap.res2);
     ++nextbyte;
-    printf("Reserved        (%02d): %02x\n", nextbyte,
+    printf("Reserved        (%02zu): %02x\n", nextbyte,
            req->message.body.tap.res3);
     ++nextbyte;
 }
@@ -254,12 +254,12 @@ static void debug_tap_mutation_request(protocol_binary_request_header *req)
 {
     debug_tap_common_body((void*)req);
     protocol_binary_request_tap_mutation *t = (void*)req;
-    int nextbyte = sizeof(protocol_binary_request_tap_no_extras);
-    printf("Item Flags   (%02d-%02d): ",
+    size_t nextbyte = sizeof(protocol_binary_request_tap_no_extras);
+    printf("Item Flags   (%02zu-%02zu): ",
            nextbyte, nextbyte + (int)sizeof(t->message.body.item.flags) - 1);
     print_uint32(t->message.body.item.flags, false, true);
     nextbyte += sizeof(t->message.body.item.flags);
-    printf("Item Expiry  (%02d-%02d): ",
+    printf("Item Expiry  (%02zu-%02zu): ",
            nextbyte, nextbyte + (int)sizeof(t->message.body.item.expiration) - 1);
     print_uint32(t->message.body.item.expiration, false, true);
     nextbyte += sizeof(t->message.body.item.expiration);
@@ -267,7 +267,7 @@ static void debug_tap_mutation_request(protocol_binary_request_header *req)
     uint8_t *ptr = (uint8_t*)(t + 1);
     uint16_t keylen = ntohs(req->request.keylen);
     if (keylen > 0) {
-        printf("Key          (%02d-%02d): [", nextbyte, nextbyte + keylen - 1);
+        printf("Key          (%02zu-%02zu): [", nextbyte, nextbyte + keylen - 1);
         print_string(ptr, (size_t)keylen);
         printf("]\n");
         ptr += keylen;
@@ -279,7 +279,7 @@ static void debug_tap_mutation_request(protocol_binary_request_header *req)
     bodylen -= req->request.extlen;
     bodylen -= ntohs(t->message.body.tap.enginespecific_length);
     if (bodylen > 0) {
-        printf("Value        (%02d-%02d): [", nextbyte, nextbyte + keylen - 1);
+        printf("Value        (%02zu-%02zu): [", nextbyte, nextbyte + keylen - 1);
         print_string(ptr, (size_t)(bodylen > 20 ? 20 : bodylen));
         printf("%s]\n", bodylen > 20 ? " ...cut..." : "");
         ptr += bodylen;
@@ -291,10 +291,10 @@ static void debug_tap_delete_request(protocol_binary_request_header *req)
 {
     protocol_binary_request_tap_no_extras *t = (void*)req;
     debug_tap_common_body(t);
-    int nextbyte = sizeof(*t);
+    size_t nextbyte = sizeof(*t);
     uint16_t keylen = ntohs(req->request.keylen);
     if (keylen > 0) {
-        printf("Key          (%02d-%02d): [", nextbyte, nextbyte + keylen - 1);
+        printf("Key          (%02zu-%02zu): [", nextbyte, nextbyte + keylen - 1);
         print_string((void*)(t + 1), (size_t)keylen);
         printf("]\n");
     }
@@ -314,8 +314,8 @@ static void debug_tap_vbucket_set_request(protocol_binary_request_header *req)
 {
     protocol_binary_request_tap_vbucket_set *t = (void*)req;
     debug_tap_common_body((void*)req);
-    int nextbyte = sizeof(protocol_binary_request_tap_no_extras);
-    printf("VB State     (%02d-%02d): ",
+    size_t nextbyte = sizeof(protocol_binary_request_tap_no_extras);
+    printf("VB State     (%02zu-%02zu): ",
            nextbyte, nextbyte + (int)sizeof(uint32_t) - 1);
     uint32_t state;
     memcpy(&state, t + 1, sizeof(state));
@@ -326,13 +326,13 @@ static void debug_tap_vbucket_set_request(protocol_binary_request_header *req)
 
 static void debug_sasl_list_mechs_response(protocol_binary_response_header *res)
 {
-    int nextbyte = sizeof(*res);
+    size_t nextbyte = sizeof(*res);
     uint32_t length = ntohl(res->response.bodylen);
     if (res->response.status == 0) {
-        printf("Mechanisms   (%02d-%02d): ",
+        printf("Mechanisms   (%02zu-%02zu): ",
                nextbyte, nextbyte + length - 1);
     } else {
-        printf("Error message(%02d-%02d): ",
+        printf("Error message(%02zu-%02zu): ",
                nextbyte, nextbyte + length - 1);
     }
     print_string((void*)(res + 1), length);
@@ -341,18 +341,18 @@ static void debug_sasl_list_mechs_response(protocol_binary_response_header *res)
 
 static void debug_sasl_auth_request(protocol_binary_request_header *req)
 {
-    int nextbyte = sizeof(*req);
+    size_t nextbyte = sizeof(*req);
     uint16_t keylen = ntohs(req->request.keylen);
     uint32_t value = ntohl(req->request.bodylen) - keylen;
 
     uint8_t *ptr = (void*)(req + 1);
-    printf("Mechanisms   (%02d-%02d): ",
+    printf("Mechanisms   (%02zu-%02zu): ",
            nextbyte, nextbyte + keylen - 1);
     print_string(ptr, keylen);
     printf("\n");
     ptr += keylen;
     nextbyte += keylen;
-    printf("Auth token   (%02d-%02d): ",
+    printf("Auth token   (%02zu-%02zu): ",
            nextbyte, nextbyte + value - 1);
     print_string(ptr, value);
     printf("\n");
@@ -360,9 +360,9 @@ static void debug_sasl_auth_request(protocol_binary_request_header *req)
 
 static void debug_sasl_auth_response(protocol_binary_response_header *res)
 {
-    int nextbyte = sizeof(*res);
+    size_t nextbyte = sizeof(*res);
     uint32_t length = ntohl(res->response.bodylen);
-    printf("Info         (%02d-%02d): ",
+    printf("Info         (%02zu-%02zu): ",
            nextbyte, nextbyte + length - 1);
     print_string((void*)(res + 1), length);
     printf("\n");
